src/main/java/net/transformatorhuis/cgi/elements/primitives/RiTorus.java
package net.transformatorhuis.cgi.elements.primitives;

import org.apache.log4j.Logger;
import net.transformatorhuis.cgi.conversion.AbstractRibType;
import net.transformatorhuis.cgi.conversion.AbstractRib;

/**
 * @author cyberroadie
 *
 * RISpec 3.2 page 83
 */
public class RiTorus extends AbstractRib {

    /**
     * Logger.
     */
    private static Logger logger = Logger.getLogger(RiTorus.class);

    /**
     * Torus.
     */
    public RiTorus() {
        super();
    }

    /**
     * Torus.
     * @param param extra parameters
     */
    public RiTorus(final String param) {
        super(param);
    }

    /**
     * Abstract method to return
     *
     * @return JAXB node returned for RIB element
     */
    public Object getJAXBNode() {
        return null;
    }

}
